[Fix] Mood cool fixes (#916)
* Fix Overlays (#756)

Overlays have a funny bug where the calls to update them are global.
Meaning if any single person gets a bad enough mood to greyscale
themselves, everyone globally gets greyscaled. This bug was also present
on Dogvision and Ultravision, and had the same cause. Frontier luckily
had a fix for those two, and the fix works here as well for the Mood
Overlay.

:cl:
- fix: Fixed an issue where Overlays(Dogvision, Ultravision, Mood) would
apply globally to all entities when updating.

* MoodSystem Crit Threshold CVar (#1010)

This PR adds a CVar that allows server hosts (such as N14 who need it)
to optionally disable the Crit Threshold modification part of the
MoodSystem. This is useful if a server has some other system that
frequently messes with Thresholds, and that a server host wishes their
system to not have to fight with Mood for it.

No changelog because this isn't playerfacing.

* Update SaturationScaleOverlay.cs

---------

Co-authored-by: VMSolidus <evilexecutive@gmail.com>