Auto Voting System (#1213)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Contains a single system, under the server-sided AutoVote folder, that
handles automatic voting of maps and presets, both of which are
configurable via a CVar.

By automatic voting system, I mean that it creates a vote for map and
preset the moment players return to lobby. If the server just started
and there are no players, it will run the vote when a player joins to
lobby.

As of now, the best use case for this is for your players to play the
modes they want, at all times.

Relevant CVars:
`vote.autovote_enabled - Is the autovote system enabled at all? (Default
is false)`
`vote.map_autovote_enabled - If it is applicable, is automatic map
voting enabled? (Default is true)`
`vote.preset_autovote_enabled - If it is applicable, is automatic preset
voting enabled? (Default is true)`

---

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Base system
- [x] CVars
- [x] Set default CVar value to false, as it is not something you might
want by default.

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added the automatic voting system.

---------

Signed-off-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>
Co-authored-by: FoxxoTrystan <45297731+FoxxoTrystan@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>