[Tweak] Mood upgrade (#931)
* Mood Upgrades

Mood Upgrades

Chemical Moodlets (And Drug Addictions!) (#896)

This PR implements a new Reagent Reaction that modifies a character's
mood, done by raising a MoodEvent on that character. It also extends
some of the functionality of the TraitSystem and MoodSystem so as to
support a new serializable Drug Addiction mechanic.

https://www.youtube.com/watch?v=8liPBsUtND4

![image](https://github.com/user-attachments/assets/3962c492-7677-4007-bf31-23e74b2b7382)

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/207210d6-e573-46e2-beb4-fab83a83d8b4)

![image](https://github.com/user-attachments/assets/4e1277e1-a873-4185-98b3-39abe06a8235)

![image](https://github.com/user-attachments/assets/e1a6cefe-82ee-482c-ace1-9fb511385fe4)

![image](https://github.com/user-attachments/assets/454dafa2-ba9e-46a6-9ac0-15e2ed35c4f8)

</p>
</details>

:cl:
- add: Drug Addictions! Drug addictions are long lasting Moodlet-Pairs.
The first consisting of a temporary Mood bonus that is refreshed by
consuming the drug, and the second consisting of an extremely long
lasting mood penalty, which replaces the mood bonus should you go a long
enough time between consuming the drug in question.
- add: Nicotine Addiction has been added as a new minor negative trait.
- add: Drugs/Reagents can now affect your character's Mood! Both with,
and without Addiction.
- add: TraitSystem has had functionality added for directly adding
Moodlets to a character upon joining the round, such as drug addictions,
or permanent mood modifications like Sanguine/Saturnine
- add: Lotophagoi Oil now induces an extremely powerful drug addiction,
providing an extremely large mood benefit for a short time. Which when
it wears off, creates an equally extreme mood penalty that lasts for a
very long time(or until you drink more Loto Oil).

Add Mood & Special Actions guidebook

Add more addictions for narcotics

* col

---------

Co-authored-by: VMSolidus <evilexecutive@gmail.com>